% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data_stand.r
\name{add_data_stand}
\alias{add_data_stand}
\title{Title}
\usage{
add_data_stand(a, df, data_type, stand_type, date, country = "spain")
}
\arguments{
\item{df}{}
}
\description{
Add individual "trees" or "saplings" data.frames to a single sf object representing a tree stand.
}
\examples{
a <- start_stand("ID1", 5, 45, "EPSG:4326")
a <- set_attributes(a)
df <- data.frame(species = c(sample(c("Pnigra","Phalep"),5,replace=T)),
dbh1 = 7.5+runif(5)*20, factor_diam1 = sample(c(127.324, 31.83099),5,replace=T))
b <- add_data_stand(a, df, "trees", "individual", 1990)

}
