% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_stands.r
\name{build_stands}
\alias{build_stands}
\title{Build tree-stand \code{sf} object}
\usage{
build_stands(a, idplot = NULL, data = list(), verbose = T)
}
\arguments{
\item{a}{a \code{sf} object containing a number of POINT geometry types.}

\item{idplot}{single identifier representing a tree stand. If it is a new plot,
a warning will be issued (unless \code{verbose=FALSE}) and a single point
will be added to \code{a}.}

\item{data}{a \code{list} with named elements. See "Details" below.}

\item{verbose}{logical, if set to TRUE warning messages may be printed.}
}
\value{
A \code{sf} object with added or modified \code{idplot} and/or fields.
}
\description{
Add or modify fields in tree-stand \code{sf} object.
}
\details{
Function \code{build_stand} must be called individually for
each tree stand.
When \code{stand_type = "individual"} and the attribute \code{country = "spain"},
the \code{df} \code{data.frame} for
trees (that is, when \code{data_type = "trees"}) must have two columns,
labeled "species" and "dbh". When it is seedlings or saplings that we want
to add to \code{a}, its format must be long, i.e. the input \code{data.frame} must have two columns, labeled
"species" and "n".
}
\examples{

# First initialize stands.
a <- start_stands()
maxdbh <- list('Pinus halepensis' = 200, 'Pinus nigra' = 230)
a <- set_parameters(a, param = list(maxdbh = maxdbh, crs =  "EPSG:4326"))

# Next, we add one stand.
df <- data.frame(species = c('Pinus halepensis', 'Quercus ilex'), dbh = c(8.6, 12.7))
a <- build_stand(a, "id1", data = list(df = df), verbose = T)

}
