% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ba.R
\name{calc_ba}
\alias{calc_ba}
\title{Calculates basal area per plot and species.}
\usage{
calc_ba(a)
}
\arguments{
\item{a}{a \code{sf} object containing a single row.}
}
\description{
It calculates the basal area, total and per species, for a plot.
}
\details{
This function is used by \code{update_stands}.
}
