% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ntrees.R
\name{calc_ntrees}
\alias{calc_ntrees}
\title{Calculates number of trees per plot and species.}
\usage{
calc_ntrees(a)
}
\arguments{
\item{a}{a \code{sf} object containing a single row.}
}
\description{
It calculates the number of trees, total and per species, for a plot.
}
\details{
This function is used by \code{update_stands}.
}
