% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_species.r
\name{calc_species}
\alias{calc_species}
\title{calc_species}
\usage{
calc_species(a)
}
\arguments{
\item{a}{\code{list} containing seedlings, saplings and trees information.}
}
\description{
Get species names for seedlings, saplings and trees in a plot \code{list}.
}
