% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stand.r
\name{check_stand}
\alias{check_stand}
\title{Check consistency of tree stands}
\usage{
check_stand(a)
}
\arguments{
\item{a}{a \code{sf} object containing a number of POINT geometry types.}
}
\value{
A \code{list} containing the number of sapling and adult tree species presents
in the stands.
}
\description{
Check consistency of tree stands
}
