% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_trees.r
\name{clear_trees}
\alias{clear_trees}
\title{Set trees to an empty \code{data.frame}}
\usage{
clear_trees(
  a,
  idplot,
  criterion = "number_trees",
  threshold = 0.01,
  per_species = T,
  update = T,
  verbose = F
)
}
\arguments{
\item{verbose}{}
}
\description{
Set trees to an empty \code{data.frame}
}
