% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diam_classes.r
\name{diam_classes}
\alias{diam_classes}
\title{Quadrature of a tabulated function at discrete intervals}
\usage{
diam_classes(x, y, xl, correction = T)
}
\arguments{
\item{x}{vector of abscissa values. It should be strictly increasing and have
the same length as \code{y}}

\item{y}{vector of ordinate values, corresponding to every value in \code{x}.}

\item{xl}{vector with limits of class intervals. Its range should be included
in the range of \code{x}.}
}
\value{
A vector containing the value of the numerical quadrature of tabulated function
\code{y} within intervals defined by \code{xl}.
}
\description{
\code{diam_classes} returns the integral of a set of points at different intervals.
}
\details{
Tabulated function (\code{x, y}) is integrated within the integration limits
given by \code{xl}. If there is an interval in \code{xl} that does not include
any value of \code{x}, the corresponding elements of the output vector will
contain a NA.
}
\examples{

# Example sinus curve.
x <- seq(-3, 60, by=.17)
y <- sin(x*pi/54)

# The exact value of this integral between 0 and 54 is (1-cos(pi))*54/pi=34.37747
xl <- seq(0, 54, length = 10)
q1 <- diam_classes(x, y, xl)
q2 <- diam_classes(x, y, xl, F)
plot((xl[-1]+xl[-length(xl)])/2, q1,xlab="Num. intervals",ylab="Integral value",pch=1)
points((xl[-1]+xl[-length(xl)])/2, q2, pch=2)
legend("topleft", pch=c(1,2), legend = c("Corrected","Uncorrected"))

# For a constant number of abscissas, the precision of the approximation degrades
as the number of intervals increases.
xl <- c(0,54)
q1 <- sum(diam_classes(x, y, xl), na.rm = T)
q2 <- sum(diam_classes(x, y, xl, F), na.rm = T)
for (i in 1:100) {
  xl <- c(0,sort(runif(i)),1)*54
  q1 <- c(q1, sum(diam_classes(x, y, xl), na.rm = T))
  q2 <- c(q2, sum(diam_classes(x, y, xl, F), na.rm = T))
}
plot(1:101, q1,ylim=c(25,35),xlab="Num. intervals",ylab="Integral value",pch=1)
points(1:101, q2, pch=2)
z <- (1-cos(pi))*54/pi
points(c(1,101), c(z,z), type = "l", lwd = 2, lty = 2)
legend("bottomleft", pch=c(1,2,NA), lty=c(NA,NA,2),
legend = c("Corrected","Uncorrected","Exact"))

}
