% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dln.R
\name{dln}
\alias{dln}
\title{The Log Normal Distribution (modified)}
\usage{
dln(x, meanlog = 0, sdlog = 1)
}
\arguments{
\item{x}{vector of quantiles.}

\item{meanlog}{mean of the distribution on the log scale.}

\item{sdlog}{standard deviation of the distribution on the log scale.}
}
\value{
Density of the log-normal distribution.
}
\description{
Density of the log-normal distribution.
}
\details{
Output is exactly the same than that of the built-in \code{stats} R function
\code{dlnorm}, but with the caveat that the case x = 0, meanlog = -Inf and sdlog = 0
now gives 0, instead of Inf as in \code{dlnorm}.

Notice that \code{log} is not given as an option, since \code{dln} is for internal
use of the package and option \code{log=F} always for our purposes.

For a full description see \code{dlnorm} help page.
}
\examples{

# Notice the difference.
dlnorm(0, -Inf, 0)
dln(0, -Inf, 0)

}
