% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrexp.r
\name{dtrexp}
\alias{dtrexp}
\title{Density of truncated exponential distribution}
\usage{
dtrexp(x, rate = 1, min = 0, max = NULL)
}
\arguments{
\item{x}{vector of quantiles.}

\item{rate}{rate of the exponential distribution.}

\item{min}{value of the lower truncation limit. Default is 0.}

\item{max}{value of the upper truncation limit. Default is infinite.}
}
\value{
density of the truncated exponential distribution.
}
\description{
Density for the truncated exponential distribution, given its rate and the lower and/or lower truncation limits.
}
\details{
Simple calculation of the density of the truncated exponential distribution.
}
\examples{

## A truncated exponential distribution.
rate <- .1
x <- seq(3,8,by=.01)
y <- dtrexp(x,rate=rate,min=3,max=8)
plot(x,y,type="l")

# Its integral must be =1 for y to be a distribution.
MiscMath::quad_ext_simpson(y,x[2]-x[1])

}
