% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_saplings.r
\name{empty_saplings}
\alias{empty_saplings}
\title{Set saplings to an empty \code{data.frame}}
\usage{
empty_saplings(a, idplot, include_date = NA)
}
\arguments{
\item{idplot}{}
}
\description{
Set saplings to an empty \code{data.frame}
}
