% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_diam_IFN.R
\name{factor_diam_IFN}
\alias{factor_diam_IFN}
\title{Areal factors to convert from IFN plot area to hectare.}
\usage{
factor_diam_IFN(x)
}
\arguments{
\item{x}{numeric, diameter at breast height in cm.}
}
\value{
A vector with the same length as \code{x} containing the conversion factor to hectare
for the Spanish Inventario Forestal Nacional.
}
\description{
Computation of areal factors for tree plots of variable radius in the
Spanish "Inventario Forestal Nacional" (IFN). Optionally, the radius of those plots
or the criteria for choosing them is given as output.
}
\details{
The diameter of tree plots of the Spanish Inventario Forestal Nacional are
sampled differently depending on their diameter at breast height (DBH).
For a description, call this function with \code{type = "interval"} and
\code{x} equal to any value.
}
\examples{
factor_diam_IFN(seq(6, 50))

}
