% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpm.R
\name{fpm}
\alias{fpm}
\title{Implementation of a forest dynamics model}
\usage{
fpm(a, data = list(), verbose = T)
}
\arguments{
\item{species}{}
}
\description{
Implementation of a forest dynamics model
}
\details{
\code{fpm} implements all steps of a full forest dynamics model, including
seedling, sapling, ingrowth and adult-tree growth and mortality processes.
}
