% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpm_growth.R
\name{fpm_growth}
\alias{fpm_growth}
\title{Calculation of tree growth}
\usage{
fpm_growth(a, type = "", data = data.frame(), models = list(), verbose = T)
}
\arguments{
\item{verbose}{}
}
\description{
It computes tree dbh growth for all species and plots.
}
