% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpm_small.R
\name{fpm_small}
\alias{fpm_small}
\title{Compute number of small trees.}
\usage{
fpm_small(a, type = "", data = data.frame(), models = list(), verbose = T)
}
\arguments{
\item{a}{a \code{sf} object containing a number of POINT geometry types.}

\item{type}{\code{character} indicating the component of the ipm model to be calculated.}

\item{data}{\code{list} whose elements are required to calculate the \code{type} component.}

\item{models}{names \code{list} with the regressions models to calculate the dynamics
of small individuals.}

\item{verbose}{\code{logical}, if set to TRUE a progress bar will be printed.}
}
\value{
A new \code{sf} stand object with the corresponding calculations.
}
\description{
It calculates the number of small trees for each class, as well as the number and
distribution of ingrowth trees.
}
\details{
This function is executed by function "fpm" and should usually not be used directly.
}
\examples{
See Vignettes.


}
