% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters.r
\name{get_parameters}
\alias{get_parameters}
\title{Extract parameters from \code{stand} object}
\usage{
get_parameters(a, param = NULL)
}
\arguments{
\item{a}{\code{sf} object containing a number of POINT geometry types}

\item{param}{character vector with names of parameters to be extracted:
\code{country}, \code{integvars}, \code{h}, \code{min_dbh}, \code{max_dbh},
\code{species} or \code{crs}. If not provided, all those parameters will
be retrieved.}
}
\value{
A \code{list} whose elements correspond to the input parameters. Parameter
\code{species} can only be computed via \code{\link{calc_species}} function.
}
\description{
Extraction of global parameters from \code{stand} object
}
\details{
See \code{\link{start_stand}} and \code{\link{set_parameters}} for details.
}
\examples{
a <- start_stands()
max_dbh <- list('Pinus halepensis' = 200, 'Pinus nigra' = 230)
a <- set_parameters(a, param = list(max_dbh = max_dbh, crs =  "EPSG:4326"))
p <- get_parameters(a)

}
