% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species.R
\name{get_species}
\alias{get_species}
\title{Extract species names of \code{sf} stand object.}
\usage{
get_species(a, verbose = T)
}
\arguments{
\item{a}{\code{sf} object containing a number of POINT geometry types}

\item{verbose}{logical, if set to TRUE a progress bar will be printed on screen.}
}
\value{
Same \code{sf} stand object with three additional fields. For parameter \code{country}="spain",
the first field, named "species", includes a list with three elements, "trees", "seedlings" and "saplings".
Each one of these
elements contains the corresponding species names per plot. The second field, named "species_all",
will be a character vector containing the names of all species in that particular plot, irrespective
of whether they correspond to trees, seedlings or saplings. The third field, named "nspecies", will
be a single number with the total number of species of trees, seedlings or saplings, present in that
plot. Therefore, this last field is simply the result of applying function \code{length} to "species_all".
A new parameter (aka attribute) is added to 'a' containing the names of all species present.

Cases other than \code{country}="spain" have not yet been implemented.
}
\description{
Extract names of all species (i.e. for Spain, seedlings, saplings and trees) per plot and stores their
names in a new field. It also adds a new parameter to 'a' called "species" containing
the names of all species present.
}
\details{
Simple counting of species per plot.
}
\examples{
a <- start_stands()
max_dbh <- list('Pinus halepensis' = 200, 'Pinus nigra' = 230)
a <- set_parameters(a, param = list(max_dbh = max_dbh, crs =  "EPSG:4326"))

# Next, we add one stand.
df <- data.frame(species = c('Pinus halepensis', 'Quercus ilex'), dbh = c(8.6, 12.7))
a <- build_stand(a, "id1", data = list(df = df), verbose = T)

# Calculate species.
a <- get_species(a)

}
