% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_stats}
\alias{get_stats}
\title{Calculate descriptive statistics of \code{sf} stands object.}
\usage{
get_stats(a, verbose = T)
}
\arguments{
\item{a}{a \code{sf} object containing a number of POINT geometry types.}

\item{verbose}{}
}
\value{
The same \code{sf} input object with new fields containing plot statistics.
}
\description{
New fields are added to the \code{stands} \code{sf} object containing plot statistics.
}
\details{
Fields are added to the \code{a} object containing the basal area per species (\code{ba_species}),
total basal area (\code{ba}), number of trees per species (\code{ntrees_species})
and total number of trees (\code{ntrees}).
}
\examples{
a <- start_stands()
max_dbh <- list('Pinus halepensis' = 200, 'Pinus nigra' = 230)
a <- set_parameters(a, param = list(max_dbh = max_dbh, crs =  "EPSG:4326"))

# Next, we add one stand.
df <- data.frame(species = c('Pinus halepensis', 'Quercus ilex'), dbh = c(8.6, 12.7))
a <- build_stands(a, "id1", data = list(df = df), verbose = T)

# Add fields with statistics.
a <- get_stats(a)

}
