% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingrowth_model.r
\name{ingrowth_model}
\alias{ingrowth_model}
\title{Title}
\usage{
ingrowth_model(obj, newdata, x, lambda, min_dbh, scale_ha = T)
}
\arguments{
\item{obj}{regression object with which to predict.}

\item{newdata}{one-row numeric \code{data.frame} containing the values of the
predictor variables.}

\item{x}{numeric vector containing the abscissas at which to calculate the
distribution of new trees.}

\item{lambda}{numeric, rate parameter for truncated exponential distribution.}

\item{min_dbh}{numeric, minimum abscissa value.}

\item{scale_ha}{logical, if set the results will be scaled up to ha.}
}
\description{
Title
}
