% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipm_ingrowth.r
\name{ipm_ingrowth}
\alias{ipm_ingrowth}
\title{Compute tree ingrowth}
\usage{
ipm_ingrowth(a, idplot, num_ingr, lambda, min_dbh, scaled_ha = T)
}
\arguments{
\item{a}{a \code{sf} object containing a number of POINT geometry types.}

\item{idplot}{identifiers of POINT elements representing tree stands to remove.}

\item{num_ingr}{named vector containing the number of new trees per species.}

\item{lambda}{named vector with the rate parameter of the truncated
exponential distribution per species.}

\item{min_dbh}{named vector containing the cutoff value for the dbh above
which an individual plant is considered an adult tree.}

\item{scaled_ha}{logical indicating whether the number of ingrowth trees
must be scaled up to hectare.}
}
\description{
Compute tree ingrowth
}
\details{
\code{ipm_ingrowth} computes the number and size distribution of new trees
}
