% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipm_spain.r
\name{ipm_spain}
\alias{ipm_spain}
\title{IPM model for Spain}
\usage{
ipm_spain(
  a,
  dat,
  reg_growth,
  reg_variance,
  reg_survival,
  reg_ingrowth,
  lambda_ingrowth,
  reg_saplings,
  quadrature = "simpson",
  progressbar = T
)
}
\arguments{
\item{a}{a \code{sf} object containing a number of POINT geometry types.}

\item{progressbar}{}

\item{idplot}{identifier of single POINT, representing a tree stand, to modify.}
}
\value{
\code{sf} object with projected adult tree population.
}
\description{
\code{ipm_quadrature} is used to calculate the integral of the continuous
distribution of adult trees times survival and diameter growth.
}
\examples{
# First initialize one single stand.
a <- start_stand("ID1", 5, 45, "EPSG:4326")
a <- set_attributes(a)

# Now we add tree information.
df <- data.frame(species = c(sample(c("Pnigra","Phalep"),5,replace=T)),
dbh = 7.5+runif(5)*20, factor_diam = sample(c(127.324, 31.83099),5,replace=T))
a <- build_stand(a, "ID1", df, "trees", "individual", 1990)

# Convolve to obtain a continuous distribution.
x <- data.frame(Pnigra = seq(7.5,200,length=100), Phalep = seq(7.5,250,length=100))
a <- set_attributes(a, integvars = x)
a <- smooth_stand(a)

# "Fake" IPM functions.
gr <- data.frame(Pnigra=rep(.1, nrow(x)), Phalep=rep(.15, nrow(x)))
va <- data.frame(Pnigra=2, Phalep=2.5)
su <- data.frame(Pnigra=rep(1, nrow(x)), Phalep=rep(1, nrow(x)))

# Apply quadrature.
b <- ipm_quadrature(a, "ID1", gr, va, su, min_dbh = 7.5)

}
