% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernsmooth.r
\name{kernsmooth}
\alias{kernsmooth}
\title{Fast kernel density estimation for a set of points}
\usage{
kernsmooth(x, y, type = "gaussian", width = NULL, normalization = T)
}
\arguments{
\item{x}{vector of abscissae values where density will be evaluated.}

\item{y}{vector with a set of points. Length of x and y need not be the same.}

\item{type}{string that determines which smoothing kernel to use.
At the moment only type = "gaussian" has been implemented.}

\item{width}{If type = "gaussian", width is the standard deviation of the
gaussian kernel.}

\item{normalization}{logical, if set the output will be normalized such that
the area below the density estimation will be equal to the number of \code{y}
points.}
}
\value{
vector with the binned kernel density estimate of points \code{y}.
}
\description{
It returns the binned kernel density estimate of a discrete set of points.
}
\details{
Density estimation takes place without the use of FFT.
}
\examples{

# Quick implementation of trapezoidal rule.
trap <- function(y, h) (sum(y) - (y[1]+y[length(y)])/2)*h

# Define points to be smoothed.
h <- .01
x <- seq(0, 100, by = h)
y <- runif(10)

# With and without normalization
znorm <- kernsmooth(x, y, width = 1)
z <- kernsmooth(x, y, width = 1, normalization = F)
print(c(With = trap(znorm, h), Without = trap(z, h)))

# Wider gaussian.
znorm <- kernsmooth(x, y, width = 10)
z <- kernsmooth(x, y, width = 10, normalization = F)
print(c(With = trap(znorm, h), Without = trap(z, h)))

plot(x, znorm, type = "l")
points(x, z, type = "l", lty = 2)

# Example with points more spread out.
y <- y*50 + 20
znorm <- kernsmooth(x, y, width = 2)
plot(x, znorm, type = "l")
points(y,rep(0, length(y)), pch = 16, cex = 1.5)

}
