% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_integvars.r
\name{make_integvars}
\alias{make_integvars}
\title{Sequence of}
\usage{
make_integvars(species, min_dbh, max_dbh, length = length)
}
\arguments{
\item{species}{character vector containing the name of species to be used.}

\item{min_dbh}{numeric vector with minimum dbh. Same length as \code{species}.}

\item{max_dbh}{numeric vector with maximum dbh. Same length as \code{species}.}

\item{length}{}
}
\value{
\code{data.frame} with as many named columns as the length of \code{species},
and with \code{length} rows. Each column corresponds to a species and it contains
an equispaced sequence of abscissas between \code{min_dbh} and \code{max_dbh}.
}
\description{
Sequence of
}
