% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxdbh.R
\docType{data}
\name{maxdbh}
\alias{maxdbh}
\title{Maximum DBH per tree species.}
\format{
An object of class \code{numeric} of length 20.
}
\usage{
data(maxdbh)
}
\description{
Maximum diameter at breast height (DBH) for twenty tree pecies present in the Spanish
Inventario Forestal Nacional. Sixteen tree species are referred to by their full
latin name, but the remaining four consist of functional groups of species and have
rather generic names.
}
\examples{
load("maxdbh")

}
\keyword{datasets}
