% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_stands.r
\name{merge_stands}
\alias{merge_stands}
\title{Merge tree stands}
\usage{
merge_stands(a, b)
}
\arguments{
\item{a}{first tree-stand \code{sf} object to merge.}

\item{b}{second tree-stand \code{sf} object to merge.}
}
\value{
A new tree-stand \code{sf} containing both all stands in \code{a} and \code{b}.
}
\description{
Merge \code{sf} objects representing tree-stands
}
\details{
Attributes in stands \code{a} and \code{b} must match for \code{merge_stand}
to work. The resulting \code{sf} object will have the same attributes as both
inputs.
}
\examples{

#' # First initialize one single stand.
a <- start_stand("ID1", 5, 45, "EPSG:4326")
a <- set_attributes(a, country = "spain")

# Next, we merge other stands.
for (i in 2:10) {
b <- start_stand(paste0("ID",i), 5, 45, "EPSG:4326")
b <- set_attributes(b, country = "spain")
a <- merge_stands(a,b)
}

}
