% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_variables.r
\name{modify_variables}
\alias{modify_variables}
\title{Title}
\usage{
modify_variables(a, idplot, df, data_type, stand_type, date)
}
\arguments{
\item{idplot}{}
}
\description{
Title
}
\examples{
a <- start_stand("ID1", 5, 45, "EPSG:4326")
a <- set_attributes(a)
df <- data.frame(species = c(sample(c("Pnigra","Phalep"),5,replace=T)),
dbh1 = 7.5+runif(5)*20, factor_diam1 = sample(c(127.324, 31.83099),5,replace=T))
b <- add_data_stand(a, df, "trees", "individual", 1990)
b <- modify_variables(b, "ID1", data_type = "saplings")

}
