% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quad_ext_simpson.r
\name{quad_ext_simpson}
\alias{quad_ext_simpson}
\title{Composite extended Simpson's quadrature rules}
\usage{
quad_ext_simpson(y, h, type = 1)
}
\arguments{
\item{y}{a numeric vector or matrix that tabulates the values of a function.
If \code{y} is matrix, the quadrature will run along rows.}

\item{h}{numeric, constant interval width.}

\item{type}{numeric, it selects algorithm 1 (\code{type=1}) from Numerical
Recipes, Eq. 4.1.14, or algorithm 2 (\code{type=2}) from
https://mathworld.wolfram.com/Newton-CotesFormulas.html.}
}
\value{
Estimated value of integral
}
\description{
It implements two composite extended Simpson's quadrature rules, as described in
"Numerical Recipes" (1989) and Wolfram' web page, to estimate the
integral of a tabulated function.
Integration is 1D, although it can applied to a vector (when input is a vector)
or along the first dimension of an input array (when input is a 2D matrix).
The latter case is faster than integrating row by row.
}
\details{
See references.
}
\examples{

## An example function.
x <- seq(2, 20, length = 9)
y <- 1-cos(x/10*pi)
quad_ext_simpson(y, x[2]-x[1], 1)
quad_ext_simpson(y, x[2]-x[1], 2)

## Compare with exact value.
exact <- (20-sin(20/10*pi)*10/pi)-(2-sin(2/10*pi)*10/pi)

}
\references{
Press, W. H., Teukolsky, S. A., Vetterling, W. T., & Flannery, B. P. (2007).
Numerical recipes 3rd edition: The art of scientific computing.
Cambridge university press.
}
