% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quad_trapez.r
\name{quad_trapez}
\alias{quad_trapez}
\title{Trapezoidal quadrature rule}
\usage{
quad_trapez(y, h)
}
\arguments{
\item{y}{a vector of ordinate values. It should have 5 or more points.}

\item{h}{interval width.}
}
\value{
estimated value of integral
}
\description{
It implements the trapezoidal quadrature rule.
Integration is 1D, although it can applied to a vector (when input is a vector) or along the first dimension
of an input array (when input is a matrix). The latter case is faster than integrating row by row.
}
\examples{

## An example function.
x <- 2:20
y <- 1-cos(x/10*pi)
quad_trapez(y,1)

## Compare with exact value.#'
(20-sin(20/10*pi)*10/pi)-(2-sin(2/10*pi)*10/pi)

## A more difficult example.
x <- 0:100
y <- x*dnorm(x,mean=0,sd=10)
quad_trapez(y,1)

## Compare with exact value.
10^2*(dnorm(min(x),mean=0,sd=10)-dnorm(max(x),mean=0,sd=10))

## Integration along first dimension of a matrix with h=0.1
y <- matrix(runif(2000),50,40)
qy <- quad_trapez(y,0.1)

}
