% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadrature.R
\name{quadrature}
\alias{quadrature}
\title{1-D composite numerical quadrature of vector or matrix.}
\usage{
quadrature(y, h = 1, type = "trapezoidal")
}
\arguments{
\item{y}{\code{numeric} vector or matrix. If matrix, the integration will take place
along rows.}

\item{h}{\code{numeric}, step size of the quadrature. It must be a positive number.}

\item{type}{\code{character}, it should be equal to "trapezoidal" (default) or "simpson"
(case insensitive).}
}
\value{
A single number (if y is a vector) and a vector (if y is a matrix) with the value of
the integration.
}
\description{
It calculates a 1-D numerical quadrature of a vector or matrix with the Composite
Trapezoidal or the Alternative Extended Simpson' rule.
}
\details{
The trapezoidal rule corresponds to the classical trapezoid formula,
which works by approximating the function by equal-width trapezoids.
\code{type} equal to "simpson" corresponds to the Alternative extended Simpson's rule
(see https://en.wikipedia.org/wiki/Simpson\%27s_rule#Composite_Simpson's_1/3_rule and
references therein).
}
\examples{
n <- 101
x <- seq(0, pi/4, len = n)
y <- cos(x)
h <- x[2]-x[1]

# Actual value of the integral of \code{cos} between 0 and pi is \code{sin(pi/4)} = 0.7071068
q1 <- c('Trapezoidal' = quadrature(y, h), 'Simpson' = quadrature(y, h, "simpson"))
my <- cbind(y, y, y)
q2 <- data.frame('Trapezoidal' = quadrature(my, h), 'Simpson' = quadrature(my, h, "simpson"))

}
