% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_stand.r
\name{remove_stand}
\alias{remove_stand}
\title{Remove tree stands}
\usage{
remove_stand(a, idplot)
}
\arguments{
\item{a}{a \code{sf} object containing a number of POINT geometry types.}

\item{idplot}{identifiers of POINT elements representing tree stands to remove.}
}
\value{
A \code{sf} object without \code{idplot} stands.
}
\description{
Remove tree stands from \code{sf} tree-stand object
}
\details{
Simple removal of rows from \code{a}.
}
\examples{

# First initialize one single stand.
a <- start_stand("ID1", 5, 45, "EPSG:4326")
a <- set_attributes(a, "spain")

# Next, we merge other stands.
for (i in 2:10) {
b <- start_stand(paste0("ID",i), 5, 45, "EPSG:4326")
b <- set_attributes(b)
a <- merge_stands(a,b)
}

# Remove two stands.
a <- remove_stand(a, c("ID3", "ID8"))

}
