% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_dataframe.R
\name{rep_dataframe}
\alias{rep_dataframe}
\title{Repeat rows in a data.frame}
\usage{
rep_dataframe(df, times)
}
\arguments{
\item{df}{a 1-row \code{data.frame} to be repeated row-wise.}

\item{times}{an integer single number specifying the number of times to repeat \code{df}.}
}
\value{
A \code{data.frame} with \code{times} rows that repeat the values in \code{df}.
}
\description{
It repeats a 1-row data.frame a number of times.
}
\details{
Little function seen in https://stackoverflow.com/questions/11121385/repeat-rows-of-a-data-frame.
}
\examples{
df <- data.frame(x = 4, y = 6)
newdf <- rep_dataframe(df, 4)

}
