% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedlings.R
\docType{data}
\name{seedlings}
\alias{seedlings}
\title{Simulated Spanish Inventario Forestal Nacional seedling data}
\format{
A data.frame with three columns containing a plot identifier "idplot",
a "n" column with an indicator of the number of seedlings present (from 0 to 3)
and a "species" column.
}
\source{
https://www.miteco.gob.es/es/biodiversidad/temas/inventarios-nacionales/inventario-forestal-nacional.html
}
\usage{
data(seedlings)
}
\description{
Simulated seedling data from 100 random plots of the Spanish Inventario Forestal Nacional.
Tree seedlings are defined as any plant with a maximum height of 30cm.
Only three species have been included: Pinus nigra, Pinus halepensis and Quercus ilex.
}
\examples{
load("seedlings")

}
\references{
Alberdi, I., Sandoval, V., Condes, S., Ca침ellas, I., & Vallejo, R. (2016).
El Inventario Forestal Nacional espa침ol, una herramienta para el conocimiento,
la gesti칩n y la conservaci칩n de los ecosistemas forestales arbolados. Ecosistemas, 25(3), 88-97.
}
\keyword{datasets}
