% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_attributes.r
\name{set_attributes}
\alias{set_attributes}
\title{Set attributes of a \code{sf} tree-stand object}
\usage{
set_attributes(a, country = NULL, integvars = NULL)
}
\arguments{
\item{a}{a \code{sf} tree-stand object created with \code{\link{start_stand}}.}

\item{country}{string indicating which country the stand belongs to.
At this moment, "Spain" (default), "France" or "USA" are valid inputs, though
only calculations for the Spanish Inventario Forestal Nacional have been
implemented. Lower or upper case letters can be used. No default value is set.}

\item{integvars}{Optional data frame containing, in columns, the abscissas for each
species to be used by the IPM methodology. Future implementations will likely
allow the use of different number of abscissas for each species, but at this
moment this is not permitted.}
}
\value{
A \code{sf} object with attributes \emph{country} and \emph{x} set.
}
\description{
Set attributes in a \code{sf} object that will be required during
calculations of tree-stand properties
}
\details{
Attribute \emph{country} must be set to either of the three possible values.
The default is "Spain", and in fact it is the only country for which
stand dynamics calculations have been implemented at this stage.
}
\examples{
a <- start_stand("ID1", 5, 45, "EPSG:4326")
a <- set_attributes(a)

}
