% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_parameters.r
\name{set_parameters}
\alias{set_parameters}
\title{Set parameters of a \code{sf} \code{stand} object}
\usage{
set_parameters(a, param = list(), verbose = T)
}
\arguments{
\item{a}{a \code{sf} stand object created with \code{\link{start_stand}}.}

\item{param}{a named \code{list} of parameters (see \code{Details} \code{\link{start_stand}}).}

\item{verbose}{logical, if set to TRUE warning messages will be shown, if required.}
}
\value{
A \code{sf} object with attributes \emph{country} and \emph{x} set.
}
\description{
Set parameters in a \code{sf} object that will be required during
calculations of stand properties
}
\details{
The elements in \code{param} input list are described in \code{\link{start_stand}}.
Notice that parameter \code{country} must be set previously in \code{\link{start_stand}}
and is ignored by \code{set_parameters} with a warning.

Setting parameter \code{integvars} automatically creates a new parameter called \code{h}
(also a list) which is the integration step per species.
}
\examples{
a <- start_stands()
max_dbh <- list('Pinus halepensis' = 200, 'Pinus nigra' = 230)
a <- set_parameters(a, param = list(max_dbh = max_dbh, crs =  "EPSG:4326"))

}
