% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_stands.r
\name{smooth_stands}
\alias{smooth_stands}
\title{Convolution of discrete tree stands}
\usage{
smooth_stands(a, smooth_type = "gaussian", width = 2, verbose = T)
}
\arguments{
\item{a}{a \code{sf} object containing a number of POINT geometry types.}

\item{smooth_type}{string indicating which smoothing window to use. Presently,
only \code{smooth_type = "gaussian"} option is available.}

\item{width}{width of smoothing window. Default is 2.}

\item{verbose}{logical, if set to TRUE a progress bar will be printed on screen.}
}
\value{
A \code{sf} object with a continuous distributions of trees per species as a
function of dbh, instead of a set of individual dbh values (as would happen
if data came from observations).
}
\description{
Convolution of discrete tree stands per species with a smoothing window.
}
\details{
Convolution with a continuous window function is done with a numerical quadrature.
The default value (\code{smooth_type = "gaussian"}) such be good enough for
most purposes.
}
\examples{

a <- start_stands()
max_dbh <- list('Pinus halepensis' = 200, 'Pinus nigra' = 230)
a <- set_parameters(a, param = list(max_dbh = max_dbh, crs =  "EPSG:4326"))

x <- list('Pinus nigra' = seq(7.5,220,length=1000),
  'Pinus halepensis' = seq(7.5,250,length=1500),
  'Quercus ilex' = seq(7.5,250,length=2000))
a <- set_parameters(a, param = list(integvars = x))

# Next, we add one stand.
df <- data.frame(species = c('Pinus halepensis', 'Quercus ilex'), dbh = c(8.6, 12.7))
a <- build_stands(a, "id1", data = list(df = df), verbose = T)

# Convolve every tree in every plot with a Gaussian window.
b <- smooth_stands(a)

}
