% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_inventory.r
\name{start_inventory}
\alias{start_inventory}
\title{Title}
\usage{
start_inventory(idplot, x, y, stand_type = NULL, date = NULL, crs = NULL)
}
\arguments{
\item{type}{}
}
\description{
Initialization of a 'sf' object.
}
\examples{
a <- start_inventory(letters[1:5],runif(5),runif(5),rep("individual",5),runif(5),"EPSG:4326")

}
