% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_stand.r
\name{start_stand}
\alias{start_stand}
\title{Start \code{sf} object to represent a tree stand}
\usage{
start_stand(idplot, x, y, crs = NA)
}
\arguments{
\item{idplot}{numeric/character vector, unique identifier of tree stands.}

\item{x}{geographic coordinate of stands (e.g. longitude if unprojected,
X in m if UTM projected)}

\item{y}{geographic coordinate of stand (same as for \code{x}).}

\item{crs}{coordinate reference system of stand. If missing, NA will be assumed.}
}
\value{
A \code{sf} object with new fields.
}
\description{
Make the basic structure of a \code{sf} object to contain stand data.
}
\details{
This simple function creates a simple feature (a.k.a. \code{sf}) POINT
object to store all data for a single tree stand.
}
\examples{
a <- start_stand("ID1", 5, 45, "EPSG:4326")

}
