% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_stands.r
\name{start_stands}
\alias{start_stands}
\title{Start \code{sf} object to represent a tree stand}
\usage{
start_stands(param = list(country = "spain"))
}
\arguments{
\item{param}{a named \code{list} of parameters (see \code{Details} below).
Default is "spain".}
}
\value{
A \code{sf} object with the following attributes (if provided in \code{param}):
\code{country}, \code{integvars}, \code{min_dbh}, \code{max_dbh}, \code{crs} and
\code{species}. New fields in the \code{sf} object are also created, depending
on which country has been selected.
}
\description{
#' @description
Create a zero-row \code{sf} object to contain stand data.
}
\details{
This simple function creates a zero-rows, simple feature (a.k.a. \code{sf})
object to store all data for a single tree stand.
The \code{param} input can supply the following components:
\itemize{
\item \code{country} string indicating which country the stand belongs to.
At this moment, "spain" (default), "france" or "usa" are valid inputs, though
only calculations for the Spanish Inventario Forestal Nacional have been
implemented. Lower or upper case letters can be used.
\item \code{integvars} optional \code{list} where each named element contains the abscissas for
the species to be modeled.
\item \code{min_dbh} named vector containing the minimum dbh after which a tree will
be considered as an adult individual.
\item \code{max_dbh} named vector containing the maximum possible dbh for each trees species.
\item \code{crs} coordinate reference system of stand. If missing, NA will be assumed.
}
}
\examples{
a <- start_stands()

}
