% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.inventory.r
\name{summary}
\alias{summary}
\title{Title}
\usage{
summary(a, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{
a <- start_inventory(letters[1:5],runif(5), runif(5),"EPSG:4326","Spain")
a <- set_type(a)
for (i in 1:5) {
df <- data.frame(species = c(sample(c("Pnigra","Phalep"),5,replace=T)),dbh = 7.5+runif(5)*20, factor_diam = sample(c(127.324, 31.83099),5,replace=T))
a <- add_stand_data(a,letters[i],df,"Spain","trees")
}
a <- update_inventory(a)
summary(a)

}
