% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\docType{data}
\name{trees}
\alias{trees}
\title{Simulated Spanish Inventario Forestal Nacional tree data}
\format{
A data.frame with five columns containing a plot identifier "idplot",
a pair of UTM coordinates for the plot, "utm_x" and "utm_y", the "dbh" of the tree (in cm)
and its corresponding "species". EPSG is 4326.
}
\usage{
data(trees)
}
\description{
Simulated adult tree data from 100 random plots of the Spanish Inventario Forestal Nacional.
Plot coordinates have been modified not to point to the actual plots.
Only three species have been selected: Pinus nigra, Pinus halepensis and Quercus ilex.
}
\examples{
load("trees")
}
\keyword{datasets}
