% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_descriptive.R
\name{calc_descriptive}
\alias{calc_descriptive}
\title{Calculates basal area per plot and species.}
\usage{
calc_descriptive(a, param = list())
}
\arguments{
\item{a}{a \code{sf} object containing a single row.}

\item{param}{param a named \code{list} of parameters (see \code{Details} below).}

\item{verbose}{}
}
\value{
The input \code{sf} object with new fields added, namely \code{ba},
\code{ntrees}, \code{ba_species} and \code{ntrees_species}, corresponding
to the total basal area (m2) and total number of trees, plus basal area and
number of trees per species, respectively.
}
\description{
It calculates the basal area and number of trees, total and per species, for a plot.
}
\details{
This function is used by \code{update_stands}. Inputs are not checked for correctness.
}
