% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_seedlings.r
\name{empty_seedlings}
\alias{empty_seedlings}
\title{Set seedlings to an empty \code{data.frame}}
\usage{
empty_seedlings(a, idplot)
}
\arguments{
\item{idplot}{}
}
\description{
Set seedlings to an empty \code{data.frame}
}
