% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saplings.R
\docType{data}
\name{saplings}
\alias{saplings}
\title{Simulated Spanish Inventario Forestal Nacional sapling data}
\format{
A data.frame with three columns containing a plot identifier "idplot",
a "n" column with the number of saplings present (from 0 to infinite)
and a "species" column.
}
\source{
https://www.miteco.gob.es/es/biodiversidad/temas/inventarios-nacionales/inventario-forestal-nacional.html
}
\usage{
data(saplings)
}
\description{
Simulated sapling data from 100 random plots of the Spanish Inventario Forestal Nacional.
Tree saplings are defined as any plant heigher than 130cm and dbh within the interval
2.5-7.5 cm.
Only three species have been included: Pinus nigra, Pinus halepensis and Quercus ilex.
}
\examples{
load("saplings")

}
\references{
Alberdi, I., Sandoval, V., Condes, S., Ca침ellas, I., & Vallejo, R. (2016).
El Inventario Forestal Nacional espa침ol, una herramienta para el conocimiento,
la gesti칩n y la conservaci칩n de los ecosistemas forestales arbolados. Ecosistemas, 25(3), 88-97.
}
\keyword{datasets}
